Moved getPipelineDescription() to the base PipelineBase class
